+++ .node-version
@@ -1,1 +1,1 @@
-v21.2.0
+v22.2.0
