+++ packages/vitest-setup/package.json
@@ -1,0 +1,7 @@
+{
+    "type": "module",
+    "dependencies": {
+        "msw": "2.3.1",
+        "core-js": "3.37.1"
+    }
+}
